more warnings

+++ decompile/General/CDSYS/CDSYS_c14_XASeek.c
@@ -8,7 +8,7 @@ void DECOMP_CDSYS_XASeek(int boolCdControl, int categoryID, int xaID)
 	if(sdata->boolUseDisc == 0) return;
 	if(sdata->bool_XnfLoaded == 0) return;
 	if(categoryID >= CDSYS_XA_NUM_TYPES) return;
-	if(xaID > DECOMP_CDSYS_XAGetNumTracks(categoryID)) return;
+	if((u_int)xaID > DECOMP_CDSYS_XAGetNumTracks(categoryID)) return;
 	
 	if(sdata->discMode != DM_AUDIO)
 		DECOMP_CDSYS_SetMode_StreamAudio();
